fix alfonso failure
